Moved FSVONavigationQueryFilterSettings to its own files
